+++ Messaging/ChangeLog.md
@@ -3,9 +3,10 @@
 ### [5.0.0] ###
 - Upgrade to .NET Standard
 - Android: Update to newer Support Libraries (v25.3.1)
-- Android: Fix support for formatting number using country's default convention via `PhoneSettings.DefaultCountryIso`. Only available for API 21 and later.
+- Android: Fix support for formatting number using country's default convention on API 21 and later. See  `PhoneSettings.DefaultCountryIso`. 
 - Android: Fix HTML formatting on Android N (API 24) and later
-- iOS: Improve `IPhoneCallTask.CanMakePhoneCall` on iOS to check support of url and connection to a carrier (#68)
+- iOS: Improve `IPhoneCallTask.CanMakePhoneCall` on iOS to check support of url and connection to a carrier [#68]
+- iOS: Fix issue with email client not appearing [#72]
 - UWP: Updated to latest UWP platform
 - **Breaking Change**: Remove Windows Store support
 - **Breaking Change**: Remove deprecated `MessagingPlugin` facade